refactored code

+++ myDemo/mysql/functions.php
@@ -1,6 +1,26 @@
 <?php include "db.php";?>
 <?php
 
+function createRecord() {
+  global $connection;
+  if (isset($_POST['submit'])) {
+
+    $username = $_POST['username'];
+    $password = $_POST['password'];
+
+    $query = "INSERT INTO users(username, password)" ;
+    $query .= "VALUES ('$username', '$password')";
+
+    $result = mysqli_query($connection, $query);
+    if (!$result) {
+      die("Query FAILED" . mysqli_error());
+    } else {
+      echo "Record Created!";
+    }
+  }
+
+}
+
 function showAllData() {
   global $connection;
   $query = "SELECT * FROM users";
@@ -17,39 +37,43 @@ function showAllData() {
 
 function updateData() {
   global $connection;
-  $username = $_POST['username'];
-  $password = $_POST['password'];
-  $id = $_POST['id'];
+  if (isset($_POST['submit'])) {
+    $username = $_POST['username'];
+    $password = $_POST['password'];
+    $id = $_POST['id'];
 
-  $query = "UPDATE users SET ";
-  $query .= "username = '$username', ";
-  $query .= "password = '$password' ";
-  $query .= "WHERE id = $id";
+    $query = "UPDATE users SET ";
+    $query .= "username = '$username', ";
+    $query .= "password = '$password' ";
+    $query .= "WHERE id = $id";
 
-  $result = mysqli_query($connection, $query);
-  if (!$result) {
-
-    die("QUERY FAILED" . mysqli_error($connection));
-
+    $result = mysqli_query($connection, $query);
+    if (!$result) {
+      die("QUERY FAILED" . mysqli_error($connection));
+    } else {
+      echo "Record Updated!";
+    }
   }
 }
 
 function deleteRecord() {
   global $connection;
-  $username = $_POST['username'];
-  $password = $_POST['password'];
-  $id = $_POST['id'];
+  if (isset($_POST['submit'])) {
+    $username = $_POST['username'];
+    $password = $_POST['password'];
+    $id = $_POST['id'];
 
-  $query = "DELETE FROM users ";
-  $query .= "WHERE id = $id";
+    $query = "DELETE FROM users ";
+    $query .= "WHERE id = $id";
 
-  $result = mysqli_query($connection, $query);
-  if (!$result) {
+    $result = mysqli_query($connection, $query);
+    if (!$result) {
 
-    die("QUERY FAILED" . mysqli_error($connection));
+      die("QUERY FAILED" . mysqli_error($connection));
 
-  } else {
-    echo "Record Deleted!";
+    } else {
+      echo "Record Deleted!";
+    }
   }
 }
